requirements.txt
python-dotenv==1.0.1
openai>=1.35.10
elevenlabs>=1.6.1
pydub>=0.25.1
ffmpeg-python>=0.2.0
Pillow>=10.4.0
tqdm>=4.66.4
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
